feat: v2 엔티티 생성
