+++ src/main/java/kr/co/hdi/dataset/repository/ProductDatasetAssignmentRepository.java
@@ -1,7 +1,7 @@
 package kr.co.hdi.dataset.repository;
 
 import kr.co.hdi.dataset.domain.ProductDatasetAssignment;
-import kr.co.hdi.user.domain.UserEntity;
+import kr.co.hdi.domain.user.entity.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
