+++ src/main/java/kr/co/hdi/dataset/service/DatasetAssignService.java
@@ -8,11 +8,11 @@ import kr.co.hdi.dataset.domain.ProductDatasetAssignment;
 import kr.co.hdi.dataset.repository.BrandDatasetAssignmentRepository;
 import kr.co.hdi.dataset.repository.BrandRepository;
 import kr.co.hdi.dataset.repository.ProductDatasetAssignmentRepository;
-import kr.co.hdi.user.domain.UserEntity;
-import kr.co.hdi.user.domain.UserType;
-import kr.co.hdi.user.exception.AuthErrorCode;
-import kr.co.hdi.user.exception.AuthException;
-import kr.co.hdi.user.repository.UserRepository;
+import kr.co.hdi.domain.user.entity.UserEntity;
+import kr.co.hdi.domain.user.entity.UserType;
+import kr.co.hdi.domain.user.exception.AuthErrorCode;
+import kr.co.hdi.domain.user.exception.AuthException;
+import kr.co.hdi.domain.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
