+++ src/main/java/kr/co/hdi/domain/data/enums/IndustryDataCategory.java
@@ -1,0 +1,7 @@
+package kr.co.hdi.domain.data.enums;
+
+public enum IndustryDataCategory {
+
+    // 2025
+    VACUUM_CLEANER, AIR_PURIFIER, HAIR_DRYER
+}
