+++ src/main/java/kr/co/hdi/domain/data/enums/VisualDataCategory.java
@@ -1,0 +1,7 @@
+package kr.co.hdi.domain.data.enums;
+
+public enum VisualDataCategory {
+
+    // 2025
+    COSMETIC, FB
+}
