+++ src/main/java/kr/co/hdi/domain/survey/entity/VisualSurvey.java
@@ -1,0 +1,34 @@
+package kr.co.hdi.domain.survey.entity;
+
+import jakarta.persistence.*;
+import kr.co.hdi.domain.survey.enums.SurveyType;
+import kr.co.hdi.domain.year.entity.AssessmentRound;
+import kr.co.hdi.global.domain.BaseTimeEntityWithDeletion;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import static jakarta.persistence.GenerationType.IDENTITY;
+import static lombok.AccessLevel.PROTECTED;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = PROTECTED)
+public class VisualSurvey extends BaseTimeEntityWithDeletion {
+
+    @Id
+    @GeneratedValue(strategy = IDENTITY)
+    @Column(name = "visual_survey_id")
+    private Long id;
+
+    private Integer surveyNumber;
+
+    private String surveyCode;
+
+    private String surveyContent;
+
+    @Enumerated(EnumType.STRING)
+    private SurveyType surveyType;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    private AssessmentRound assessmentRound;
+}
