+++ src/main/java/kr/co/hdi/domain/survey/enums/SurveyType.java
@@ -1,0 +1,5 @@
+package kr.co.hdi.domain.survey.enums;
+
+public enum SurveyType {
+    NUMBER, TEXT
+}
