+++ src/main/java/kr/co/hdi/domain/user/dto/RegisterRequest.java
@@ -1,0 +1,13 @@
+package kr.co.hdi.domain.user.dto;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record RegisterRequest(
+        @NotBlank
+        String email,
+        @NotBlank
+        String password,
+        @NotBlank
+        String name
+) {
+}
