+++ src/main/java/kr/co/hdi/domain/user/dto/request/LoginRequest.java
@@ -1,0 +1,11 @@
+package kr.co.hdi.domain.user.dto.request;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record LoginRequest(
+        @NotBlank
+        String email,
+        @NotBlank
+        String password
+) {
+}
