+++ src/main/java/kr/co/hdi/domain/user/dto/response/AuthInfoResponse.java
@@ -1,0 +1,9 @@
+package kr.co.hdi.domain.user.dto.response;
+
+import kr.co.hdi.domain.user.entity.UserType;
+
+public record AuthInfoResponse(
+        UserType userType,
+        Boolean surveyDone
+) {
+}
