+++ src/main/java/kr/co/hdi/domain/user/dto/response/AuthResponse.java
@@ -1,0 +1,32 @@
+package kr.co.hdi.domain.user.dto.response;
+
+
+import kr.co.hdi.domain.user.entity.Role;
+import kr.co.hdi.domain.user.entity.UserType;
+import kr.co.hdi.domain.user.entity.UserEntity;
+
+public record AuthResponse(
+        Long id,
+        String email,
+        String name,
+        Role role,
+        UserType userType,
+        Boolean surveyDone
+) {
+
+    public static AuthResponse from(UserEntity user
+    ) {
+        return new AuthResponse(
+                user.getId(),
+                user.getEmail(),
+                user.getName(),
+                user.getRole(),
+                user.getUserType(),
+                user.getSurveyDone()
+        );
+    }
+
+    public static AuthResponse of(Long userId, String email, String name, Role role, UserType userType, Boolean surveyDone) {
+        return new AuthResponse(userId, email, name, role, userType, surveyDone);
+    }
+}
