+++ src/main/java/kr/co/hdi/domain/user/entity/Role.java
@@ -1,0 +1,12 @@
+package kr.co.hdi.domain.user.entity;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+@Getter
+public enum Role {
+    USER("ROLE_USER"), ADMIN("ROLE_ADMIN");
+
+    private final String value;
+}
