+++ src/main/java/kr/co/hdi/domain/user/entity/UserType.java
@@ -1,0 +1,7 @@
+package kr.co.hdi.domain.user.entity;
+
+public enum UserType {
+
+    BRAND, PRODUCT,
+    VISUAL, INDUSTRY
+}
