+++ src/main/java/kr/co/hdi/domain/user/exception/AuthException.java
@@ -1,0 +1,14 @@
+package kr.co.hdi.domain.user.exception;
+
+
+import kr.co.hdi.global.exception.CustomException;
+
+public class AuthException extends CustomException {
+    public AuthException(AuthErrorCode errorCode) {
+        super(errorCode);
+    }
+
+    public AuthException(AuthErrorCode errorCode, String message) {
+        super(errorCode, message);
+    }
+}
