+++ src/main/java/kr/co/hdi/domain/user/repository/UserRepository.java
@@ -1,0 +1,14 @@
+package kr.co.hdi.domain.user.repository;
+
+import kr.co.hdi.domain.user.entity.UserEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
+    Optional<UserEntity> findByEmail(String email);
+
+    Optional<UserEntity> findByName(String name);
+
+    boolean existsByEmail(String email);
+}
