+++ src/main/java/kr/co/hdi/domain/year/entity/Year.java
@@ -1,0 +1,24 @@
+package kr.co.hdi.domain.year.entity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import kr.co.hdi.global.domain.BaseTimeEntityWithDeletion;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import static jakarta.persistence.GenerationType.IDENTITY;
+import static lombok.AccessLevel.PROTECTED;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = PROTECTED)
+public class Year extends BaseTimeEntityWithDeletion {
+
+    @Id @GeneratedValue(strategy = IDENTITY)
+    @Column(name = "year_id")
+    private Long id;
+
+    private Integer year;
+}
