+++ src/main/java/kr/co/hdi/domain/year/enums/DomainType.java
@@ -1,0 +1,7 @@
+package kr.co.hdi.domain.year.enums;
+
+public enum DomainType {
+
+    VISUAL,
+    INDUSTRY
+}
