+++ src/main/java/kr/co/hdi/result/service/ResultService.java
@@ -6,10 +6,10 @@ import kr.co.hdi.survey.domain.WeightedScore;
 import kr.co.hdi.survey.repository.BrandResponseRepository;
 import kr.co.hdi.survey.repository.ProductResponseRepository;
 import kr.co.hdi.survey.repository.WeightedScoreRepository;
-import kr.co.hdi.user.domain.UserEntity;
-import kr.co.hdi.user.exception.AuthErrorCode;
-import kr.co.hdi.user.exception.AuthException;
-import kr.co.hdi.user.repository.UserRepository;
+import kr.co.hdi.domain.user.entity.UserEntity;
+import kr.co.hdi.domain.user.exception.AuthErrorCode;
+import kr.co.hdi.domain.user.exception.AuthException;
+import kr.co.hdi.domain.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.poi.ss.usermodel.Row;
