+++ src/main/java/kr/co/hdi/survey/domain/WeightedScore.java
@@ -2,7 +2,7 @@ package kr.co.hdi.survey.domain;
 
 import jakarta.persistence.*;
 import kr.co.hdi.global.domain.BaseTimeEntityWithDeletion;
-import kr.co.hdi.user.domain.UserEntity;
+import kr.co.hdi.domain.user.entity.UserEntity;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
