+++ src/main/java/kr/co/hdi/survey/repository/BrandResponseRepository.java
@@ -2,7 +2,7 @@ package kr.co.hdi.survey.repository;
 
 import kr.co.hdi.survey.domain.BrandResponse;
 import kr.co.hdi.survey.domain.ResponseStatus;
-import kr.co.hdi.user.domain.UserEntity;
+import kr.co.hdi.domain.user.entity.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
