+++ src/main/java/kr/co/hdi/survey/repository/WeightedScoreRepository.java
@@ -1,7 +1,7 @@
 package kr.co.hdi.survey.repository;
 
 import kr.co.hdi.survey.domain.WeightedScore;
-import kr.co.hdi.user.domain.UserEntity;
+import kr.co.hdi.domain.user.entity.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
