+++ src/main/java/kr/co/hdi/survey/service/SurveyService.java
@@ -12,11 +12,11 @@ import kr.co.hdi.survey.dto.request.WeightedScoreRequest;
 import kr.co.hdi.survey.exception.SurveyErrorCode;
 import kr.co.hdi.survey.exception.SurveyException;
 import kr.co.hdi.survey.repository.*;
-import kr.co.hdi.user.domain.UserEntity;
-import kr.co.hdi.user.domain.UserType;
-import kr.co.hdi.user.exception.AuthErrorCode;
-import kr.co.hdi.user.exception.AuthException;
-import kr.co.hdi.user.repository.UserRepository;
+import kr.co.hdi.domain.user.entity.UserEntity;
+import kr.co.hdi.domain.user.entity.UserType;
+import kr.co.hdi.domain.user.exception.AuthErrorCode;
+import kr.co.hdi.domain.user.exception.AuthException;
+import kr.co.hdi.domain.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
