--- src/main/java/kr/co/hdi/user/domain/Role.java
@@ -1,12 +1,0 @@
-package kr.co.hdi.user.domain;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-@Getter
-public enum Role {
-    USER("ROLE_USER"), ADMIN("ROLE_ADMIN");
-
-    private final String value;
-}
