--- src/main/java/kr/co/hdi/user/domain/UserType.java
@@ -1,6 +1,0 @@
-package kr.co.hdi.user.domain;
-
-public enum UserType {
-
-    BRAND, PRODUCT
-}
