--- src/main/java/kr/co/hdi/user/dto/RegisterRequest.java
@@ -1,13 +1,0 @@
-package kr.co.hdi.user.dto;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record RegisterRequest(
-        @NotBlank
-        String email,
-        @NotBlank
-        String password,
-        @NotBlank
-        String name
-) {
-}
