--- src/main/java/kr/co/hdi/user/dto/request/LoginRequest.java
@@ -1,11 +1,0 @@
-package kr.co.hdi.user.dto.request;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record LoginRequest(
-        @NotBlank
-        String email,
-        @NotBlank
-        String password
-) {
-}
