--- src/main/java/kr/co/hdi/user/dto/response/AuthInfoResponse.java
@@ -1,9 +1,0 @@
-package kr.co.hdi.user.dto.response;
-
-import kr.co.hdi.user.domain.UserType;
-
-public record AuthInfoResponse(
-        UserType userType,
-        Boolean surveyDone
-) {
-}
