--- src/main/java/kr/co/hdi/user/exception/AuthErrorCode.java
@@ -1,18 +1,0 @@
-package kr.co.hdi.user.exception;
-
-import kr.co.hdi.global.exception.ErrorCode;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-
-@RequiredArgsConstructor
-@Getter
-public enum AuthErrorCode implements ErrorCode {
-
-    USER_NOT_FOUND(HttpStatus.NOT_FOUND, "사용자를 찾을 수 없습니다."),
-    INVALID_PASSWORD(HttpStatus.UNAUTHORIZED, "비밀번호가 올바르지 않습니다."),
-    USER_ALREADY_EXISTS(HttpStatus.BAD_REQUEST, "이미 존재하는 이메일입니다.");
-
-    private final HttpStatus httpStatus;
-    private final String message;
-}
